Undo for Place, Drop, Pickup; Associated cleanup
To make undoing sensible, dropping all stones from the carry stack in the movement state
doesn't revert the state back to Start. If this was kept in, the history action version of drop
would have to potentially remember the movement state it was in previously.
This functionality will now be done by the client instead, using the undo tools.