Added estimation of game alloc size
Uses deepsize, behind the feature flag "measure_game_size". Prints the total estimated size after each move, for the total game, and which part of that is the "game state" (everything except the game history).
The estimation seems mostly accurate, maybe a few bytes off (there seems to be some weirdness with Stone allocation in stone stacks).
The purpose of this was mostly to assess how big the game state is, to decide whether history scrolling can just give out clones of the game state, or should have one internal state that it hands out a reference to.